Draft custom error handler to control reject_on_worker_lost
