+++ src/conf/celeryconf_v2.py
@@ -37,6 +37,7 @@ CELERYD_PREFETCH_MULTIPLIER = 1
 # setup queues so that tasks aren't removed from the queue until
 # complete and reschedule if the task worker goes offline
 CELERY_ACKS_LATE = True
+#CELERY_REJECT_ON_WORKER_LOST = True
 CELERY_REJECT_ON_WORKER_LOST = False
 CELERY_TASK_QUEUES = (Broadcast('model-worker-broadcast'), )
