+++ src/model_execution_worker/distributed_tasks.py
@@ -35,10 +35,9 @@ from ..conf.iniconf import settings
 from .backends.aws_storage import AwsObjectStore
 from .backends.azure_storage import AzureObjectStore
 from .storage_manager import BaseStorageConnector
+from .celery_request_handler import FailureWorkerRetry
 
 
-
-from billiard.einfo import ExceptionWithTraceback
 # testing
 import random
 
@@ -51,46 +50,8 @@ ARCHIVE_FILE_SUFFIX = 'tar.gz'
 RUNNING_TASK_STATUS = OASIS_TASK_STATUS["running"]["id"]
 TASK_LOG_DIR = settings.get('worker', 'TASK_LOG_DIR', fallback='/var/log/oasis/tasks')
 
-#celery_conf.worker_task_kwargs['autoretry_for'] = (Exception, WorkerLostError)
-
-class CustomRequest(CeleryRequest):
-    """ Using 'reject_on_worker_lost=True' to re-queue a task on `WorkerLostError` results
-        in an infinite loop if the task OOM errors each try.
-
-        so instead this overrides 'on_failure' to call the standard `on_retry` call like if an
-        exception occurs
-
-        src: https://docs.celeryq.dev/en/stable/_modules/celery/worker/request.html#Request
-    """
-    def on_failure(self, exc_info, send_failed_event=True, return_ok=False):
-        #@from celery.contrib import rdb; rdb.set_trace()
-        exc = exc_info.exception
-
-        #if isinstance(exc, ExceptionWithTraceback):
-        #    exc = exc.exc
-
-        if isinstance(exc, WorkerLostError):
-            return self.on_retry(exc_info)
-            exc_info = Retry(message='worker lost')
-            exc_info.exception = exc
-            ##return self.retry(exc=exc)
-
-
-        super().on_failure(exc_info, send_failed_event, return_ok)  # Call the base class method
-
-# Monkey patch
-#celery.worker.request.Request = CustomErrorHandler
-
-class FailureWorkerRetry(Task):
-    Request = CustomRequest
-
 app = Celery(task_cls=FailureWorkerRetry)
 app.config_from_object(celery_conf)
-
-
-
-#app.request = CustomErrorHandler
-#from celery.contrib import rdb; rdb.set_trace()
 
 logging.info(dir(app))
 
@@ -734,6 +695,7 @@ def prepare_keys_file_chunk(
         )
 
         # mimic chunk lost
+        os._exit(1)
         if random.randrange(0,3):
             logging.info('--- Keys chunk lost ---')
             os._exit(1)
@@ -1206,7 +1168,6 @@ def handle_task_failure(*args, sender=None, task_id=None, **kwargs):
         if not keep_remote_data:
             logging.info(f"deleting remote data, {dir_remote_data}")
             filestore.delete_dir(dir_remote_data)
-
 
 @before_task_publish.connect
 def mark_task_as_queued_receiver(*args, headers=None, body=None, **kwargs):
